Moves RouteTransition to shared folder
