--- app/components/TabbedPage/TabContent.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const TabContent = ({
-  children
-}) => (
-  <div className="tab-content-wrapper">
-    {children}
-  </div>
-);
-
-export default TabContent;
