+++ app/components/shared/index.js
@@ -1,1 +1,2 @@
+export { default as RouteTransition } from "./RouteTransition";
 export { default as Tooltip } from "./Tooltip";
