+++ app/style/Layout.less
@@ -61,13 +61,6 @@
   bottom: 0px;
 }
 
-.tab-content-wrapper {
-  position: absolute;
-  left: 60px;
-  right: 60px;
-  top: 60px;
-}
-
 .tab-content-card {
   background-color: white;
 }
